+++ example/client.py
@@ -1,5 +1,7 @@
 #!/usr/bin/env python
 
+# WS client example
+
 import asyncio
 import websockets
 
@@ -7,10 +9,11 @@ async def hello():
     async with websockets.connect(
             'ws://localhost:8765') as websocket:
         name = input("What's your name? ")
+
         await websocket.send(name)
-        print("> {}".format(name))
+        print(f"> {name}")
 
         greeting = await websocket.recv()
-        print("< {}".format(greeting))
+        print(f"< {greeting}")
 
 asyncio.get_event_loop().run_until_complete(hello())