+++ example/secure_server.py
@@ -9,11 +9,12 @@ import websockets
 
 async def hello(websocket, path):
     name = await websocket.recv()
-    print("< {}".format(name))
+    print(f"< {name}")
 
-    greeting = "Hello {}!".format(name)
+    greeting = f"Hello {name}!"
+
     await websocket.send(greeting)
-    print("> {}".format(greeting))
+    print(f"> {greeting}")
 
 ssl_context = ssl.SSLContext(ssl.PROTOCOL_TLS_SERVER)
 ssl_context.load_cert_chain(