+++ docs/changelog.rst
@@ -29,6 +29,8 @@ Also:
 * If a :meth:`~protocol.WebSocketCommonProtocol.ping` doesn't receive a pong,
   it's cancelled when the connection is closed.
 
+* Updated documentation with new features from Python 3.6.
+
 * Fixed missing close code, which caused :exc:`TypeError` on connection close.
 
 4.0