+++ docs/changelog.rst
@@ -18,6 +18,8 @@ Changelog
 
 Also:
 
+* Added compatibility with Python 3.7.
+
 * :func:`~client.connect()` performs HTTP Basic Auth when the URI contains
   credentials.
 
@@ -125,6 +127,8 @@ Also:
 
 3.3
 ...
+
+* Ensured compatibility with Python 3.6.
 
 * Reduced noise in logs caused by connection resets.